preliminary support for emitting HTML anchors for function defns
Eventually, I hope to add an HTML anchor for every function, datatype,
and class defined in a module, so that it is possible for haddock to
link directly to the appropriate part of the source code.  For now, this
patch just lays the groundwork, and does the simplest possible recognition
of a function definition.  Doesn't cover infix defns, doesn't ignore
type signatures - just the absolute basics.

darcs-hash:20060125183655-6871e-828586494470c601a2c6b6decd744b37d551c33c