Now -nopartial is the default behaviour with literate files.
Previously, it was not possible to get the HTML / CSS / LaTeX prologue
and epilogue in conjunction with the -lit option.  Now, the top'n'tail
is used by default, and you must explicitly ask for -partial if you
want the previous behaviour.

darcs-hash:20090824120208-6871e-ed99b144f05c8749d925a8822dd6d45b30748239